Create generate package.
Move PasswordGenerator to org.passay.generate package.
Add FillRemainingCharactersAppender implementation to have all character appending concerns as appender implementations.
